requirements.txt
# Core RPA
pyautogui==0.9.54
pyperclip==1.8.2
pytesseract==0.3.10
pyscreeze==0.1.29

# Image Processing
opencv-python==4.8.0
pillow==10.0.0
numpy==1.24.0

# Data Processing
pandas==2.1.0
openpyxl==3.1.0

# Utilities
schedule==1.2.0
python-dotenv==1.0.0
psutil==5.9.0
colorama==0.4.6
tqdm==4.66.1
requests==2.31.0

# Development
pytest==7.4.2